requirements.txt
pandas==2.2.2
xgboost==2.1.0
lightgbm==4.4.0
scikit-learn==1.5.1
PyYAML==6.0.1
optuna==3.6.1
hierarchicalforecast==0.4.1
numpy==1.23.5
statsforecast==1.7.6
bayesian-optimization